bitcoin cat lion haircut
A long-haired cat that has just gotten a lion haircut.  On the side of the cat's body is an intricately shaved bitcoin symbol in the cat's hair.  The cat groomer probably spent hours to carefully cut each of the cat's hairs on its side to create such an elaborate bitcoin design.  It's truly a work of art.